app/requirements.txt
flask
flask-cors
mysql-connector-python
pytest
flake8
